Move DefaultRewardFromObservation into spaces namespace.
Move DefaultRewardFromObservation out of compiler_env.py and into the
compiler_gym.spaces namespace. This is just housekeeping to minimize
the number of classes and definitions in compiler_env.py.
